+++ src/sixtypical/compiler.py
@@ -1,7 +1,7 @@
 # encoding: UTF-8
 
 from sixtypical.ast import (
-    Program, Routine, Block, SingleOp, Call, GoTo, If, Repeat, For, WithInterruptsOff, Save, PointInto
+    Program, Routine, Block, SingleOp, Reset, Call, GoTo, If, Repeat, For, WithInterruptsOff, Save, PointInto
 )
 from sixtypical.model import (
     ConstantRef, LocationRef, IndexedRef, IndirectRef,
@@ -34,16 +34,17 @@ class Compiler(object):
         self.symtab = symtab
         self.emitter = emitter
         self.routines = {}           # routine.name -> Routine
-        self.routine_statics = {}    # routine.name -> { static.name -> Label }
+        self.routine_locals = {}     # routine.name -> { local.name -> Label }
         self.labels = {}             # global.name -> Label  ("global" includes routines)
         self.trampolines = {}        # Location -> Label
+        self.pointer_assoc = {}      # pointer name -> table name  (I'm not entirely happy about this)
         self.current_routine = None
 
     # - - - - helper methods - - - -
 
     def get_type_for_name(self, name):
-        if self.current_routine and self.symtab.has_static(self.current_routine.name, name):
-            return self.symtab.fetch_static_type(self.current_routine.name, name)
+        if self.current_routine and self.symtab.has_local(self.current_routine.name, name):
+            return self.symtab.fetch_local_type(self.current_routine.name, name)
         return self.symtab.fetch_global_type(name)
 
     def get_type(self, ref):
@@ -76,9 +77,9 @@ class Compiler(object):
 
     def get_label(self, name):
         if self.current_routine:
-            static_label = self.routine_statics.get(self.current_routine.name, {}).get(name)
-            if static_label:
-                return static_label
+            local_label = self.routine_locals.get(self.current_routine.name, {}).get(name)
+            if local_label:
+                return local_label
         return self.labels[name]
 
     def absolute_or_zero_page(self, label):
@@ -107,16 +108,15 @@ class Compiler(object):
                 label.set_addr(routine.addr)
             self.labels[routine.name] = label
 
-            if hasattr(routine, 'statics'):
-                self.current_routine = routine
-                static_labels = {}
-                for defn in routine.statics:
-                    length = self.compute_length_of_defn(defn)
-                    label = Label(defn.name, addr=defn.addr, length=length)
-                    static_labels[defn.name] = label
-                    declarations.append((defn, self.symtab.fetch_static_type(routine.name, defn.name), label))
-                self.routine_statics[routine.name] = static_labels
-                self.current_routine = None
+            self.current_routine = routine
+            local_labels = {}
+            for defn in routine.locals:
+                length = self.compute_length_of_defn(defn)
+                label = Label(defn.name, addr=defn.addr, length=length)
+                local_labels[defn.name] = label
+                declarations.append((defn, self.symtab.fetch_local_type(routine.name, defn.name), label))
+            self.routine_locals[routine.name] = local_labels
+            self.current_routine = None
 
         if compilation_roster is None:
             compilation_roster = [['main']] + [[routine.name] for routine in program.routines if routine.name != 'main']
@@ -192,6 +192,8 @@ class Compiler(object):
             return self.compile_save(instr)
         elif isinstance(instr, PointInto):
             return self.compile_point_into(instr)
+        elif isinstance(instr, Reset):
+            return self.compile_reset(instr)
         else:
             raise NotImplementedError
 
@@ -742,12 +744,16 @@ class Compiler(object):
                 self.emitter.emit(STA(Absolute(src_label)))
 
     def compile_point_into(self, instr):
-        src_label = self.get_label(instr.table.name)
+        self.pointer_assoc[instr.pointer.name] = instr.table.name
+        self.compile_block(instr.block)
+        del self.pointer_assoc[instr.pointer.name]
+
+    def compile_reset(self, instr):
+        table_name = self.pointer_assoc[instr.pointer.name]
+        src_label = Offset(self.get_label(table_name), instr.offset.value)
         dest_label = self.get_label(instr.pointer.name)
 
         self.emitter.emit(LDA(Immediate(HighAddressByte(src_label))))
         self.emitter.emit(STA(ZeroPage(dest_label)))
         self.emitter.emit(LDA(Immediate(LowAddressByte(src_label))))
         self.emitter.emit(STA(ZeroPage(Offset(dest_label, 1))))
-
-        self.compile_block(instr.block)